flask/requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
circuitbreaker==1.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cryptography==42.0.7
cx-Oracle==8.3.0
decorator==4.4.2
deep-translator==1.11.4
filelock==3.13.1
Flask==3.0.3
flatbuffers==24.3.25
fsspec==2024.2.0
huggingface-hub==0.23.1
humanfriendly==10.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
JPype1==1.5.0
konlpy==0.6.0
libaio==0.9.3
llvmlite==0.42.0
lxml==5.2.2
MarkupSafe==2.1.5
mkl==2021.4.0
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
nudenet==3.0.8
numba==0.59.1
numpy==1.26.4
oci==2.126.4
onnxruntime==1.18.0
openai-whisper==20231117
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
proglog==0.1.10
protobuf==5.27.0
pycparser==2.22
pyOpenSSL==24.1.0
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.2
safetensors==0.4.3
six==1.16.0
soupsieve==2.5
sympy==1.12
tbb==2021.11.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.41.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
